fixup fontconfig.rb (linux mint 17, gcc 4.9)
After brew updates, some of the issues in #164 are solved,
the only remaining issue is the docs.

This commit disables compiling the docs, and instead tries
to install the pregenerated documentation that comes with
the tarball. Without a patch like this, my compilation stops
with errors when trying to generate the docs.

See compilation notes here:
http://www.linuxfromscratch.org/blfs/view/svn/general/fontconfig.html